SCRUM-70: add curly brackets where missing, remove unnecessary casts and whitelines
git-svn-id: file:///home/migrator/subversion-migration/svn-repo-copy/community/magnolia/trunk@44895 a15fbcf9-bee4-0310-bafe-d02d26ad3b5f

+++ magnolia-core/src/main/java/info/magnolia/module/model/Version.java
@@ -266,9 +266,9 @@ public class Version {
     @Override
     public int hashCode() {
         int result;
-        result = (int) major;
-        result = 31 * result + (int) minor;
-        result = 31 * result + (int) patch;
+        result = major;
+        result = 31 * result + minor;
+        result = 31 * result + patch;
         result = 31 * result + (classifier != null ? classifier.hashCode() : 0);
         return result;
     }
